+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 Merged a little from alex into multistage.
 Added else clause to RestrictedMiddleUniform.
 Fixed plot-point-eating bug.
+Added configuration option `memory_report_period`.
 
 ### 2023-02-23
 New pattern RestrictedMiddleUniform.