+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## next [0.5.4]
+## [0.5.4]
 
 ### 2023-02-24
 Merged a little from alex into multistage.
@@ -8,6 +8,7 @@ Added else clause to RestrictedMiddleUniform.
 Fixed plot-point-eating bug.
 Added configuration option `memory_report_period`.
 Fixed vc of injection.
+Publish 0.5.4.
 
 ### 2023-02-23
 New pattern RestrictedMiddleUniform.